Added ActivitySource as default approach for activity generating
If we didn't set ActivitySource, it will be set to string.Empty by default. It means that our activity belongs to the namespace string.Empty.
In our case we need to set namespace for OpenTelemetry, because it subscribes to activity sources by name. But this 'scoping' feature can be useful for other purposes as well